Core dynamic scaling functional
The core functionality of the dynamic scaling is complete. With this change, the immediate children of the focal object will be scaled up, in order to reduce empty space, as well as make the objects more visible. Upon zooming in on an object, it and the other children are scaled down, and the new immediate children are scaled up. These scaling operations are animated to be smooth transitions, occurring in parallel to the zoom animation.