Moved version tag & load parent model on zoom
Adjusted the location of the version tag to be "wrapped" around the top-left corner of the navbar, at a 45deg angle. Required setting navbars `overflow: hidden`.

Included the loading of the parent object model when zooming to an object. This will usually only be used when the site is opened on a specific planet/satellite, instead of the sun.