Fixed mouse coordinates issue
Mouse coordinate calculations were incorrect when the pinned container was hidden off to the right side of the window, due to the container expanding the inner width of the window. To prevent this, the smallest value between inner (body of the website) and outer (entire window size) dimensions are used.

Also fixed a bug where content would overlap with the navbar if the window height was small enough. Content is now fixed to `4rem` from the top.